\name{plot_weights}
\alias{plot_weights}
\title{Plot Robust Weights}
\usage{
plot_weights(object, ...)
}
\arguments{
  \item{object}{A fitted \code{robustcause_rlm} or \code{robustcause_s} object.}
  \item{...}{Additional graphical parameters passed to \code{plot}.}
}
\description{
Plots per-observation robustness weights from a fitted model.
}
\seealso{
\code{\link{vcov_robust}}, \code{\link{confint_robust}}
}
